+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v0.1.9
+- Added directory index file listings
+
 ## v0.1.8
 - Improved handling of program exits
 - Fixed memory leak w/ deflateText calling delete instead of delete[]